# openapi.yaml
openapi: 3.0.3
info:
  title: Anna's Archive API
  description: |
    A RESTful API to search and retrieve book metadata from Anna's Archive.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BookSummary:
      type: object
      properties:
        md5:
          type: string
          description: 32-character MD5 hash
        title:
          type: string
          nullable: true
        author:
          type: string
          nullable: true
        publisher:
          type: string
          nullable: true
        year:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        extension:
          type: string
          nullable: true
        filesize:
          type: string
          nullable: true
          description: Human-readable filesize (e.g. "0.6 MB")
        description:
          type: string
          nullable: true
        cover:
          type: string
          nullable: true
          description: Cover image URL
        filename:
          type: string
          nullable: true
        downloads:
          type: integer
          nullable: true
        lists:
          type: integer
          nullable: true
        bookType:
          type: string
          nullable: true
          description: Content category (e.g. "Book (fiction)")
        hasIssues:
          type: boolean
        url:
          type: string
          description: Full URL to the book detail page
    BookDetail:
      type: object
      properties:
        md5:
          type: string
          description: 32-character MD5 hash
        title:
          type: string
          nullable: true
        author:
          type: string
          nullable: true
        publisher:
          type: string
          nullable: true
        year:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        extension:
          type: string
          nullable: true
        filesize:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        cover:
          type: string
          nullable: true
        collections:
          type: array
          items:
            type: string
        contentType:
          type: string
          nullable: true
        zlibId:
          type: string
          nullable: true
        isbn:
          type: array
          items:
            type: string
        ipfsCids:
          type: array
          items:
            type: string
        filepaths:
          type: array
          items:
            type: string
        url:
          type: string
        downloadLinks:
          $ref: '#/components/schemas/BookDownloadLinks'
        metadata:
          type: object
          description: Variable key-value metadata bag. Keys are dynamic (e.g. sha256, asin, series, dates, identifiers, etc.)
          additionalProperties: true
    BookResponse:
      type: object
      properties:
        success:
          type: boolean
        md5:
          type: string
        book:
          $ref: '#/components/schemas/BookDetail'
        domain:
          type: string
        cached:
          type: boolean
        responseTime:
          type: integer
...
    LinkStatus:
      type: object
      properties:
        label:
          type: string
          description: Display label for the link
        url:
          type: string
          description: The download URL
    BookDownloadLinks:
      type: object
      properties:
        fast:
          type: array
          items:
            $ref: '#/components/schemas/LinkStatus'
        slow:
          type: array
          items:
            $ref: '#/components/schemas/LinkStatus'
        external:
          type: array
          items:
            $ref: '#/components/schemas/LinkStatus'
paths:
...
  /api/search:
    get:
      tags: [Search]
      summary: Search for books
      description: Search Anna's Archive with filters and advanced search options.
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          description: Search query string (optional — blank searches are allowed)
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: lang
          in: query
          description: Language filter (e.g., "en")
          schema:
            type: string
        - name: ext
          in: query
          description: File extension filter (e.g., "pdf")
          schema:
            type: string
        - name: sort
          in: query
          description: Sort order for results
          schema:
            type: string
            enum: [most_relevant, newest, oldest, largest, smallest, newest_added, oldest_added, random]
            default: most_relevant
        - name: content
          in: query
          description: Content type filter
          schema:
            type: string
            enum: [book_any, book_fiction, book_nonfiction, magazine, standards_document, comics, other]
        - name: index
          in: query
          description: Source index filter
          schema:
            type: string
            enum: [journals, digital_lending, meta]
        - name: termtype_1
          in: query
          description: First advanced search field type
          schema:
            type: string
            enum: [title, author, publisher, edition_varia, year, original_filename, description_comments]
        - name: termval_1
          in: query
          description: First advanced search field value
          schema:
            type: string
        - name: termtype_2
          in: query
          description: Second advanced search field type
          schema:
            type: string
            enum: [title, author, publisher, edition_varia, year, original_filename, description_comments]
        - name: termval_2
          in: query
          description: Second advanced search field value
          schema:
            type: string
        - name: termtype_3
          in: query
          description: Third advanced search field type
          schema:
            type: string
            enum: [title, author, publisher, edition_varia, year, original_filename, description_comments]
        - name: termval_3
          in: query
          description: Third advanced search field value
          schema:
            type: string
        - name: termtype_4
          in: query
          description: Fourth advanced search field type
          schema:
            type: string
            enum: [title, author, publisher, edition_varia, year, original_filename, description_comments]
        - name: termval_4
          in: query
          description: Fourth advanced search field value
          schema:
            type: string
        - name: termtype_5
          in: query
          description: Fifth advanced search field type
          schema:
            type: string
            enum: [title, author, publisher, edition_varia, year, original_filename, description_comments]
        - name: termval_5
          in: query
          description: Fifth advanced search field value
          schema:
            type: string
        - name: termtype_6
          in: query
          description: Sixth advanced search field type
          schema:
            type: string
            enum: [title, author, publisher, edition_varia, year, original_filename, description_comments]
        - name: termval_6
          in: query
          description: Sixth advanced search field value
          schema:
            type: string
        - name: termtype_7
          in: query
          description: Seventh advanced search field type
          schema:
            type: string
            enum: [title, author, publisher, edition_varia, year, original_filename, description_comments]
        - name: termval_7
          in: query
          description: Seventh advanced search field value
          schema:
            type: string
        - name: termtype_8
          in: query
          description: Eighth advanced search field type
          schema:
            type: string
            enum: [title, author, publisher, edition_varia, year, original_filename, description_comments]
        - name: termval_8
          in: query
          description: Eighth advanced search field value
          schema:
            type: string
        - name: termtype_9
          in: query
          description: Ninth advanced search field type
          schema:
            type: string
            enum: [title, author, publisher, edition_varia, year, original_filename, description_comments]
        - name: termval_9
          in: query
          description: Ninth advanced search field value
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  query:
                    type: string
                  page:
                    type: integer
                  count:
                    type: integer
                  filters:
                    type: object
                    description: Active filters (including advanced search fields if provided)
                  results:
                    type: object
                    properties:
                      primary:
                        type: array
                        items:
                          $ref: '#/components/schemas/BookSummary'
                      partial:
                        type: array
                        items:
                          $ref: '#/components/schemas/BookSummary'
                  cached:
                    type: boolean
                  responseTime:
                    type: integer
  /api/book/{md5}:
    get:
      tags: [Book]
      summary: Get book details
      description: Fetch full metadata and download links for a specific book.
      security:
        - BearerAuth: []
      parameters:
        - name: md5
          in: path
          required: true
          description: 32-character MD5 hash of the book
          schema:
            type: string
            pattern: '^[a-f0-9]{32}$'
        - name: refresh
          in: query
          description: Bypass cache and force a fresh scrape
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Book details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '404':
          description: Book not found
...
  /api/cache:
    delete:
      tags: [System]
      summary: Flush API cache
      description: Admin only. Clear all cached search and book data.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Cache cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string